feat: integrate native Renoise CLI workflows
Delegate plugin commands to the native CLI with a bundled legacy fallback and shared credential resolution.

Add universal setup and storyboard skills, adapter tests, and CI coverage.
