+++ .claude-plugin/marketplace.json
@@ -7,7 +7,7 @@
   "plugins": [
     {
       "name": "renoise",
-      "version": "1.0.1",
+      "version": "1.1.0",
       "source": "./",
       "description": "AI video production skills — creative direction, generation, editing, and e-commerce content"
     }
