+++ .claude-plugin/plugin.json
@@ -1,7 +1,7 @@
 {
   "name": "renoise",
   "description": "AI video production skills — creative direction, generation, editing, and e-commerce content",
-  "version": "1.0.1",
+  "version": "1.1.0",
   "author": {
     "name": "Renoise"
   },
