+++ .github/workflows/test.yml
@@ -1,0 +1,22 @@
+name: Test
+
+on:
+  push:
+  pull_request:
+
+permissions:
+  contents: read
+
+jobs:
+  adapter:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/setup-node@v4
+        with:
+          node-version: 22
+      - run: npm test
+      - run: node --check skills/renoise-gen/renoise-cli.mjs
+      - run: node --check skills/renoise-gen/renoise-cli-legacy.mjs
+      - run: bash -n hooks/session-start.sh skills/director/scripts/batch-generate.sh
+      - run: npm pack --dry-run
