+++ openclaw.plugin.json
@@ -2,7 +2,7 @@
   "id": "renoise",
   "name": "Renoise",
   "description": "AI video production skills — creative direction, generation, analysis, e-commerce content, and download",
-  "version": "1.0.1",
+  "version": "1.1.0",
   "configSchema": {
     "type": "object",
     "additionalProperties": false,
@@ -12,6 +12,8 @@
     "skills/director",
     "skills/gemini-gen",
     "skills/renoise-gen",
+    "skills/renoise-setup",
+    "skills/storyboard-sheet",
     "skills/video-download"
   ]
 }
