+++ package.json
@@ -1,9 +1,12 @@
 {
   "name": "@renoise/plugin",
-  "version": "1.0.1",
+  "version": "1.1.0",
   "license": "Apache-2.0",
   "type": "module",
   "description": "AI video production skills — creative direction, generation, editing, and e-commerce content",
+  "scripts": {
+    "test": "node --test tests/*.test.mjs"
+  },
   "repository": {
     "type": "git",
     "url": "https://github.com/ArcoCodes/renoise-plugins-official.git"
@@ -19,6 +22,7 @@
     "hooks/",
     "commands/",
     "skills/",
+    "src/",
     "README.md"
   ]
 }
