+++ src/credits-cache.ts
@@ -7,6 +7,10 @@
 import fs from 'fs'
 import path from 'path'
 import os from 'os'
+import { execFile } from 'child_process'
+import { promisify } from 'util'
+
+const execFileAsync = promisify(execFile)
 
 const CACHE_DIR = path.join(os.homedir(), '.renoise')
 const CACHE_FILE = path.join(CACHE_DIR, 'credits-cache.json')
@@ -76,7 +80,18 @@ export function getCredits(): { data: CreditsData | null; fresh: boolean } {
  */
 export async function refreshFromApi(): Promise<void> {
   const apiKey = process.env.RENOISE_API_KEY
-  if (!apiKey) return
+  if (!apiKey) {
+    try {
+      const cli = process.env.RENOISE_CLI_PATH || 'renoise'
+      const { stdout } = await execFileAsync(cli, ['account', 'status', '--json'], { encoding: 'utf8', timeout: 5000 })
+      const json = JSON.parse(stdout) as { credit?: { balance?: number } }
+      const balance = json?.credit?.balance
+      if (typeof balance === 'number') writeCache(balance, 'renoise')
+    } catch {
+      // Native CLI is optional; the plugin adapter still has its bundled fallback.
+    }
+    return
+  }
 
   const baseUrl = process.env.RENOISE_BASE_URL || 'https://www.renoise.ai/api/public/v1'
   const headers: Record<string, string> = { 'X-API-Key': apiKey }
